+++ front/src/Socket.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import toast from 'react-hot-toast';
 import { SetterOrUpdater, useSetRecoilState } from 'recoil';
-import { IChartItem, IDailyLog } from '@src/types';
+import { IChartItem, IDailyLog, IStockExecutionInfo, IStockExecutionItem } from '@src/types';
 import chartAtom from '@recoil/chart';
 import HoldStockListAtom from '@recoil/holdStockList';
 import dailyLogAtom from '@recoil/stockDailyLog';
+import stockExecutionAtom from '@recoil/stockExecution';
 import fetchHoldStocks from '@common/utils/fetchHoldStocks';
 import webSocketAtom from '@recoil/websocket/atom';
 import stockListAtom, { IStockListItem, IStockChartItem } from '@recoil/stockList/atom';
 import { IAskOrderItem, IBidOrderItem, askOrdersAtom, bidOrdersAtom } from '@recoil/stockOrders/index';
-import stockExecutionAtom, { IStockExecutionInfo, IStockExecutionItem } from './recoil/stockExecution/atom';
 import { translateRequestData, translateResponseData } from './common/utils/socketUtils';
 import Emitter from './common/utils/eventEmitter';
