+++ front/src/recoil/stockExecution/index.ts
@@ -1,6 +1,1 @@
-import atom from './atom';
-import type { IStockExecutionItem } from './atom';
-
-export { IStockExecutionItem };
-
-export default atom;
+export { default } from './atom';
